#418 REFACTOR Generate new github client code and fix code to mathc changes

+++ code/src/githubc2/githubc2_components_authorization.ml
@@ -22,7 +22,7 @@ module Primary = struct
     expires_at : string option;
     fingerprint : string option;
     hashed_token : string option;
-    id : int;
+    id : int64;
     installation : Githubc2_components_nullable_scoped_installation.t option; [@default None]
     note : string option;
     note_url : string option;
