+++ code/src/githubc2/githubc2_components_pull_request_review.ml
@@ -35,7 +35,7 @@ module Primary = struct
     body_text : string option; [@default None]
     commit_id : string option;
     html_url : string;
-    id : int;
+    id : int64;
     node_id : string;
     pull_request_url : string;
     state : string;
